+++ tests/VObject/Property/ICalendar/DateTimeTest.php
@@ -270,7 +270,7 @@ class DateTimeTest extends \PHPUnit_Framework_TestCase {
     }
 
     /**
-     * @expectedException LogicException
+     * @expectedException \Sabre\VObject\InvalidDataException
      */
     function testGetDateTimeDateInvalid() {