+++ lib/Recur/EventIterator.php
@@ -71,6 +71,21 @@ class EventIterator implements \Iterator {
      */
     protected $allDay = false;
 
+
+    /**
+     * Automatically stop iterating after this many iterations.
+     *
+     * This is a security measure. Without this, it would be possible to craft
+     * specific events that recur many, many times, potentially DDOSing the
+     * server.
+     *
+     * The default (3500) allows creation of a dialy event that goes on for 10
+     * years, which is hopefully long enouogh for most.
+     *
+     * Set this value to -1 to disable this control altogether.
+     */
+    static $maxInstances = 3500;
+
     /**
      * Creates the iterator.
      *
@@ -314,6 +329,9 @@ class EventIterator implements \Iterator {
      */
     function valid() {
 
+        if ($this->counter > self::$maxInstances && self::$maxInstances !== -1) {
+            throw new MaxInstancesExceededException('Recurring events are only allowed to generate ' . self::$maxInstances);
+        }
         return !!$this->currentDate;
 
     }