+++ lib/InvalidDataException.php
@@ -3,11 +3,12 @@
 namespace Sabre\VObject;
 
 /**
- * Exception thrown by Reader if an invalid object was attempted to be parsed.
+ * This exception is thrown whenever an invalid value is found anywhere in a
+ * iCalendar or vCard object.
  *
  * @copyright Copyright (C) 2011-2015 fruux GmbH (https://fruux.com/).
  * @author Evert Pot (http://evertpot.com/)
  * @license http://sabre.io/license/ Modified BSD License
  */
-class ParseException extends InvalidDataException {
+class InvalidDataException extends \Exception {
 }