+++ tests/VObject/Recur/EventIterator/Issue26Test.php
@@ -1,11 +1,14 @@
 <?php
 
-namespace Sabre\VObject;
+namespace Sabre\VObject\Recur\EventIterator;
+
+use Sabre\VObject\Reader;
+use Sabre\VObject\Recur\EventIterator;
 
 class Issue26Test extends \PHPUnit_Framework_TestCase {
 
     /**
-     * @expectedException \InvalidArgumentException
+     * @expectedException \Sabre\VObject\InvalidDataException
      */
     function testExpand() {
 
@@ -24,7 +27,7 @@ ICS;
         $vcal = Reader::read($input);
         $this->assertInstanceOf('Sabre\\VObject\\Component\\VCalendar', $vcal);
 
-        $it = new Recur\EventIterator($vcal, 'bae5d57a98');
+        $it = new EventIterator($vcal, 'bae5d57a98');
         iterator_to_array($it);
 
     }